Architecture updates
