+++ tests/test_fa/test_fa_excel.py
@@ -1,10 +1,10 @@
-""" fundamental_analysis/dcf_model.py tests """
+""" fundamental_analysis/dcf_view.py tests """
 from unittest import TestCase
 
 import vcr
 
-from gamestonk_terminal.stocks.fundamental_analysis.dcf_model import CreateExcelFA
-import gamestonk_terminal.stocks.fundamental_analysis.excel.variables as var
+from gamestonk_terminal.stocks.fundamental_analysis.dcf_view import CreateExcelFA
+import gamestonk_terminal.stocks.fundamental_analysis.dcf_model as md
 
 
 class TestExcelClass(TestCase):
@@ -12,17 +12,17 @@ class TestExcelClass(TestCase):
         "tests/cassettes/test_fa/test_fa_excel/test_covers_all_tickers.yaml"
     )
     def test_covers_all_tickers(self):
-        for ticker in var.tickers:
+        for ticker in md.tickers:
             excel = CreateExcelFA(ticker, False)
             df_is = excel.get_data("IS", 1, True)
-            items_is = var.non_gaap_is + var.gaap_is
+            items_is = md.non_gaap_is + md.gaap_is
             for item in df_is.index:
                 self.assertIn(item, items_is)
             df_bs = excel.get_data("BS", 1, True)
-            items_bs = var.non_gaap_bs + var.gaap_bs
+            items_bs = md.non_gaap_bs + md.gaap_bs
             for item in df_bs.index:
                 self.assertIn(item, items_bs)
             df_cf = excel.get_data("CF", 1, True)
-            items_cf = var.non_gaap_cf + var.gaap_cf
+            items_cf = md.non_gaap_cf + md.gaap_cf
             for item in df_cf.index:
                 self.assertIn(item, items_cf)
